Refactor RTT targeting system
As or Irrlicht 1.8.4, material type binding to 2D stuff is not supported for OpenGL, so GUI is now rendered to a separate RTT. Also, due to an undefined behaviour for 'setRenderTarget' when the previous RTT is an array of 'IRenderTarget', common RTT such as the scene one and the GUI one are moved to the 'SharedData' class, though their texture manipulations are still performed in the 'Engine' class. The scene shader is also modified to mix correctly scene and GUI textures on the screen.